+++ src/lib.rs
@@ -727,10 +727,6 @@ pub mod routing;
 pub mod service;
 pub mod sse;
 
-#[cfg(feature = "ws")]
-#[cfg_attr(docsrs, doc(cfg(feature = "ws")))]
-pub mod ws;
-
 #[cfg(test)]
 mod tests;